structures - width

+++ Printf/libftprintf/ft_itoa_base.c
@@ -6,7 +6,7 @@
 /*   By: tbillon <tbillon@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/07 10:19:11 by tbillon           #+#    #+#             */
-/*   Updated: 2020/12/07 10:27:49 by tbillon          ###   ########lyon.fr   */
+/*   Updated: 2020/12/07 13:05:11 by tbillon          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_itoa_base(long int n, char *base)
 {
-	char				*str;
-	unsigned long long	i;
+	char	*str;
+	long	i;
 
 	i = ft_strlen(base);
 	if (!(str = (char *)malloc(sizeof(char) * 2)))
